refactor : repository 분리
2021. 08. 23.
domain 안에 있던 repository 분리

Resolves : loop-study
See also : sah3122
