+++ src/main/java/kitchenpos/menu/respository/JpaMenuRepository.java
@@ -1,5 +1,6 @@
-package kitchenpos.menu.domain;
+package kitchenpos.menu.respository;
 
+import kitchenpos.menu.domain.Menu;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
