+++ src/main/java/kitchenpos/order/repository/OrderRepository.java
@@ -1,5 +1,7 @@
-package kitchenpos.order.domain;
+package kitchenpos.order.repository;
 
+import kitchenpos.order.domain.Order;
+import kitchenpos.order.domain.OrderStatus;
 import kitchenpos.ordertable.domain.OrderTable;
 
 import java.util.List;
