+++ src/main/java/kitchenpos/product/respository/JpaProductRepository.java
@@ -1,7 +1,6 @@
-package kitchenpos.product.domain;
+package kitchenpos.product.respository;
 
 import kitchenpos.product.domain.Product;
-import kitchenpos.product.domain.ProductRepository;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.UUID;
