+++ src/test/java/kitchenpos/menugroup/domain/MenuGroupServiceTest.java
@@ -1,8 +1,7 @@
 package kitchenpos.menugroup.domain;
 
-import kitchenpos.menugroup.domain.InMemoryMenuGroupRepository;
-import kitchenpos.menugroup.domain.MenuGroup;
-import kitchenpos.menugroup.domain.MenuGroupRepository;
+import kitchenpos.menugroup.repository.InMemoryMenuGroupRepository;
+import kitchenpos.menugroup.repository.MenuGroupRepository;
 import kitchenpos.menugroup.application.MenuGroupService;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
