+++ src/test/java/kitchenpos/order/domain/OrderServiceTest.java
@@ -1,12 +1,14 @@
 package kitchenpos.order.domain;
 
-import kitchenpos.ordertable.domain.InMemoryOrderTableRepository;
+import kitchenpos.order.repository.InMemoryOrderRepository;
+import kitchenpos.order.repository.OrderRepository;
+import kitchenpos.ordertable.repository.InMemoryOrderTableRepository;
 import kitchenpos.infra.FakeKitchenridersClient;
-import kitchenpos.menu.domain.InMemoryMenuRepository;
-import kitchenpos.menu.domain.MenuRepository;
+import kitchenpos.menu.repository.InMemoryMenuRepository;
+import kitchenpos.menu.respository.MenuRepository;
 import kitchenpos.order.application.OrderService;
 import kitchenpos.ordertable.domain.OrderTable;
-import kitchenpos.ordertable.domain.OrderTableRepository;
+import kitchenpos.ordertable.ordertable.OrderTableRepository;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
