+++ src/test/java/kitchenpos/order/repository/InMemoryOrderRepository.java
@@ -1,8 +1,8 @@
-package kitchenpos.order.domain;
+package kitchenpos.order.repository;
 
 import kitchenpos.order.domain.Order;
-import kitchenpos.order.domain.OrderRepository;
 import kitchenpos.order.domain.OrderStatus;
+import kitchenpos.order.repository.OrderRepository;
 import kitchenpos.ordertable.domain.OrderTable;
 
 import java.util.*;
